src/main/java/school/faang/user_service/controller/SkillController.java
package school.faang.user_service.controller;

import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Component;
import school.faang.user_service.dto.SkillCandidateDto;
import school.faang.user_service.dto.SkillDto;
import school.faang.user_service.service.SkillService;

import java.util.List;
import java.util.Optional;

@Component
@RequiredArgsConstructor
public class SkillController {
    private final SkillService skillService;

    public SkillDto create(SkillDto skill) {
        if(validateSkill(skill)) {
            return skillService.create(skill);
        } else throw new IllegalArgumentException("Имя не должно быть пустым.");
    }

    public boolean validateSkill(SkillDto skill) {
        return skill.getTitle() != null && !skill.getTitle().isEmpty();
    }

    public List<SkillDto> getUserSkills(long userId) {
        return skillService.getUserSkills(userId);
    }

    public List<SkillCandidateDto> getOfferedSkills(long userId) {
        return skillService.getOfferedSkills(userId);
    }

    public Optional<SkillDto> acquireSkillFromOffers(long skillId, long userId) {
        return skillService.acquireSkillFromOffers(skillId, userId);
    }
}
